render door and event

+++ src/org/vafada/daxviewer/GeoWallRecord.java
@@ -1,5 +1,5 @@
 package org.vafada.daxviewer;
 
-public record GeoWallRecord(int row, int col, int north, int south, int east, int west, byte event, int door) {
+public record GeoWallRecord(int row, int col, int north, int south, int east, int west, int event, int door) {
 
 }
